chore: add basic query.tz tests
