+++ package.json
@@ -34,6 +34,7 @@
     "echarts": "^5.3.0",
     "jsdom": "^19.0.0",
     "moment": "^2.29.1",
+    "moment-timezone": "^0.5.34",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
     "rxjs": "^7.2.0",
