+++ src/dto/base/quality.dto.ts
