+++ src/dto/get-theme.query.dto.ts
@@ -1,5 +1,5 @@
 import { Mixin } from 'ts-mixer';
-import { OutputFormatDto } from './output-format.dto';
-import { QualityDto } from './quality.dto';
+import { OutputFormatDto } from './base/output-format.dto';
+import { QualityDto } from './base/quality.dto';
 
 export class GetThemeQueryDto extends Mixin(OutputFormatDto, QualityDto) {}
