+++ src/types/theme.enum.ts
@@ -1,0 +1,7 @@
+export enum PresetTheme {
+  GREEN = 'green',
+  RED = 'red',
+  PURPLE = 'purple',
+  BLUE = 'blue',
+  YELLOW = 'yellow',
+}
